feat(parity): Major parity improvements across Player, Team, INI, W3D systems
Player class (player.rs):
- Added ~45 new fields matching C++ Player class
- Implemented init(), update(), xfer(), crc() methods
- Added PlayerRelationMap integration
- Implemented radar, battle plans, sciences, rank/skill point systems

Team class (team.rs):
- Added TeamTemplateInfo struct with 34 fields
- Implemented TCreateUnitsInfo struct
- Added AttitudeType enum
- Implemented update_state(), kill_team(), evacuate_team()
- Added trigger area detection methods
- Implemented complete xfer() for save/load
- Added TeamPrototype methods

INI Parsers:
- Added 5 new parsers: InGameUI, CommandMap, HeaderTemplate, ScriptAction, ScriptCondition
- Fixed GameLOD wrapper functions for type conversion
- Added missing block parser registrations

W3D/Rendering:
- Enhanced shadow system (ShadowType, W3DShadowManager)
- Fixed texture manager mipmap/HSV algorithms
- Improved default draw callback parity

Other fixes:
- RAMFile class complete implementation
- StreamingArchiveFile File trait implementation
- Money class xfer() fix
- Weapon defaults, radius damage cone, shared reload time

+++ GeneralsRust/Code/GameEngine/Common/src/common/rts/production_prerequisite.rs
@@ -604,7 +604,7 @@ impl Snapshotable for ProductionPrerequisite {
     }
 
     fn xfer(&mut self, xfer: &mut dyn Xfer) -> Result<(), String> {
-        const CURRENT_VERSION: u8 = 1;
+        const CURRENT_VERSION: u32 = 1;
         let mut version = CURRENT_VERSION;
         xfer.xfer_version(&mut version, CURRENT_VERSION)
             .map_err(|e| e.to_string())?;
